remover/requirements.txt
albucore==0.0.16
albumentations==1.4.15
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
dnspython==2.7.0
easydict==1.13
email_validator==2.2.0
eval_type_backport==0.2.2
fastapi-cli==0.0.7
filelock==3.16.1
flatbuffers==24.12.23
flet==0.25.2
fsspec==2024.12.0
gdown==5.2.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
imageio==2.36.1
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kornia==0.8.0
kornia_rs==0.1.8
lazy_loader==0.4
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
onnxruntime==1.20.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pika==1.3.2
pillow==11.1.0
pip3-autoremove==1.2.2
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
PyMatting==1.1.13
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
referencing==0.36.1
rembg==2.0.61
repath==0.9.0
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
safetensors==0.5.2
scikit-image==0.25.0
scipy==1.15.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
tifffile==2025.1.10
timm==1.0.13
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transparent-background==1.3.3
triton==3.1.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.1
wget==3.2
